refactor: move getDataForEmail file into no-meteor folder
