+++ imports/plugins/core/accounts/server/no-meteor/mutations/inviteShopMember.js
@@ -3,7 +3,7 @@ import SimpleSchema from "simpl-schema";
 import Random from "@reactioncommerce/random";
 import ReactionError from "@reactioncommerce/reaction-error";
 import getCurrentUserName from "/imports/plugins/core/accounts/server/no-meteor/util/getCurrentUserName";
-import getDataForEmail from "/imports/plugins/core/accounts/server/util/getDataForEmail";
+import getDataForEmail from "/imports/plugins/core/accounts/server/no-meteor/util/getDataForEmail";
 
 const inputSchema = new SimpleSchema({
   email: String,
