+++ imports/plugins/core/accounts/server/no-meteor/util/getDataForEmail.js
